Clean up rendering logic between share and website mode
