+++ share/templates/send.html
@@ -28,24 +28,27 @@
       <th id="size-header">Size</th>
       <th></th>
     </tr>
-    {% for info in file_info.dirs %}
+    {% for info in dirs %}
     <tr>
       <td>
         <img width="30" height="30" title="" alt="" src="{{ static_url_path }}/img/web_folder.png" />
-        {{ info.basename }}
+        <a href="{{ info.basename }}">
+          {{ info.basename }}
+        </a>
       </td>
-      <td>{{ info.size_human }}</td>
-      <td></td>
+      <td>&mdash;</td>
     </tr>
     {% endfor %}
-    {% for info in file_info.files %}
+
+    {% for info in files %}
     <tr>
       <td>
         <img width="30" height="30" title="" alt="" src="{{ static_url_path }}/img/web_file.png" />
-        {{ info.basename }}
+        <a href="{{ info.basename }}">
+          {{ info.basename }}
+        </a>
       </td>
       <td>{{ info.size_human }}</td>
-      <td></td>
     </tr>
     {% endfor %}
   </table>
